refactor: split digest.go and agent_linux_state.go into focused files
internal/report/digest.go (1106 LOC -> 646 + 296 + 193)
  - digest_types.go: row structs, DigestInput, and small standalone
    utilities (TruncateExeForDigest, TruncateUTF8ToMaxBytes, BPFCmdName).
  - digest_aggregation.go: hot-egress ranking, mode classification,
    ringbuf summing, truthfulness interpretation, KPI visibility
    predicates, and per-channel empty-state reason helpers.
  - digest.go: markdown writers (writeTriageRibbon / writeHotEgressTable
    / writeDetectProfileKPI) plus BuildDetectMarkdown and the
    WriteDetectDigest entry point. Adds a package-level doc comment
    explaining the three-file layout.

internal/agent/agent_linux_state.go (1029 LOC -> 548 + 347 + 167)
  - agent_linux_state_sections.go: per-section state (fork / fs /
    network), the BPF wire-format constants, ring-read retry backoff,
    integrity canary, and the small fixed-cap row/edge buffers.
  - agent_linux_state_enforce.go: enforcement state, LSM-vs-cgroup
    backend selection, and the typed enforceDenyError used to fail-fast.
  - agent_linux_state.go: keeps runStats (cumulative counters), its
    locked accessors, snapshotSummary, and rowBuffer + trimRing — they
    stay together because rowBuffer feeds dropped-event accounting back
    into runStats. Adds a header comment describing the split.

Pure code motion: no signature or behaviour changes. go build / gofmt
clean; full `go test ./...` (23 packages) passes.
